implemented dorm dropdown when assigning room

+++ dormTrack_backend/src/main/java/com/cboard/dormTrack/dormTrack_backend/repository/DormRepository.java
@@ -3,4 +3,4 @@ package com.cboard.dormTrack.dormTrack_backend.repository;
 import com.cboard.dormTrack.dormTrack_backend.model.Dorm;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DormRepository extends JpaRepository<Dorm, Integer> {}
+public interface DormRepository extends JpaRepository<Dorm, Long> {}
